Only delete memory in ~PoolBuffer if not finalizing pools
